Remove redundant parameters from Facilitator
I updated the Facilitator's constructor and method to exclude
unnecessary arguments.

The AbiBinProvider is now injected into the Facilitator so that the
Facilitator can provide the ABI to the StakeHelper. A longer, proper fix
outside of the scope of this issue would be to update the StakeHelper to
not require the ABI.

I updated the README. If I understood the example correctly, there was a
racing condition for the bounty amount which I fixed by adding `await`
inside an `async` function.

It was not possible to inject the stake helper instead of instantiating
it inside the Facilitator. Redundant data is carried over too far and
fixing the StakeHelper is not in the scope of this issue.

Fixes #89